==> 4-find_best_model.ipynb <==
import pandas as pd
import numpy as np
import requests as rq
import matplotlib.pyplot as plt
from libraries.Cryptov2 import Cryptocurrency
from arch.univariate import arch_model

df_BTC = pd.read_csv("./models/ts/fits/BTC-results.csv", index_col=0)
df_EOS = pd.read_csv("./models/ts/fits/EOS-results.csv", index_col=0)
df_ETH = pd.read_csv("./models/ts/fits/ETH-results.csv", index_col=0)
df_IOTA = pd.read_csv("./models/ts/fits/IOTA-results.csv", index_col=0)

# Model selection
## According to the Akaike selection criteria

df_BTC[df_BTC['AIC_score'] == min(df_BTC['AIC_score'])]

df_EOS[df_EOS['AIC_score'] == min(df_EOS['AIC_score'])]

df_ETH[df_ETH['AIC_score'] == min(df_ETH['AIC_score'])]

df_IOTA[df_IOTA['AIC_score'] == min(df_IOTA['AIC_score'])]

round(df_BTC.sort_values(by='AIC_score', ascending=True)[['volatility_model','dist','p','q','AIC_score','BIC_score','LL_score']][:5],2).to_csv('./models/ts/fits/BTC-ordered-AIC-results.csv')
round(df_EOS.sort_values(by='AIC_score', ascending=True)[['volatility_model','dist','p','q','AIC_score','BIC_score','LL_score']][:5],2).to_csv('./models/ts/fits/EOS-ordered-AIC-results.csv')
round(df_ETH.sort_values(by='AIC_score', ascending=True)[['volatility_model','dist','p','q','AIC_score','BIC_score','LL_score']][:5],2).to_csv('./models/ts/fits/ETH-ordered-AIC-results.csv')
round(df_IOTA.sort_values(by='AIC_score', ascending=True)[['volatility_model','dist','p','q','AIC_score','BIC_score','LL_score']][:5],2).to_csv('./models/ts/fits/IOTA-ordered-AIC-results.csv')

# Model selection
## According to the BIC selection criteria

df_BTC[df_BTC['BIC_score'] == min(df_BTC['BIC_score'])]

df_EOS[df_EOS['BIC_score'] == min(df_EOS['BIC_score'])]

df_ETH[df_ETH['BIC_score'] == min(df_ETH['BIC_score'])]

df_IOTA[df_IOTA['BIC_score'] == min(df_IOTA['BIC_score'])]

# Model selection
## According to the maximization of the log likelihood

df_BTC[df_BTC['LL_score'] == max(df_BTC['LL_score'])]

df_EOS[df_EOS['LL_score'] == max(df_EOS['LL_score'])]

df_ETH[df_ETH['LL_score'] == max(df_ETH['LL_score'])]

df_IOTA[df_IOTA['LL_score'] == max(df_IOTA['LL_score'])]

# Model selection
## Accortding to Model Confidence Set
1. BTC : Model 56
1. EOS : Model 73
1. IOTA : Model 8
1. ETH : Model 73

crypto_list = ['BTC','EOS','ETH','IOTA']

df_BTC.iloc[56,:]

df_ETH.iloc[73,:]

df_IOTA.iloc[8,:]

df_EOS.iloc[73,:]
